+++ crates/pdk-api/src/macros.rs
@@ -4,7 +4,7 @@
 macro_rules! config_struct {
     ($struct:item) => {
         #[derive(Debug, serde::Deserialize)]
-        #[serde(default, deny_unknown_fields, rename_all = "camelCase")]
+        #[serde(default, rename_all = "camelCase")]
         $struct
     };
 }